[process] Divide ps command and collect non-truncated wchan as a new one.
wchan column is now truncated because there are many columns.
It is impossible to distinguish the symbol name if there are many
similar symbol names.

Example: here is an output of ps command when plymouthd process
got hanged:

  F  UID    PID PPID PRI  NI    VSZ   RSS WCHAN  STAT TTY      TIME COMMAND
  5    0    159    1  20   0  44644  1696 n_tty_ S    ?        0:04 /bin/plymouthd --attach-to-session

The prefix n_tty_ is used in tty device driver of linux kernel
as follows:

  struct tty_ldisc_ops tty_ldisc_N_TTY = {
      .magic           = TTY_LDISC_MAGIC,
      .name            = "n_tty",
      .open            = n_tty_open,
      .close           = n_tty_close,
      .flush_buffer    = n_tty_flush_buffer,
      .chars_in_buffer = n_tty_chars_in_buffer,
      .read            = n_tty_read,
      .write           = n_tty_write,
      .ioctl           = n_tty_ioctl,
      .set_termios     = n_tty_set_termios,
      .poll            = n_tty_poll,
      .receive_buf     = n_tty_receive_buf,
      .write_wakeup    = n_tty_write_wakeup,
      .fasync          = n_tty_fasync,
  };

There are three ideas to fix this issue:

  1) Specify the best size of wchan.
  2) Use nwchan instead of wchan.
  3) Divide ps command and collect wchan with the maximum kernel symbol size by new one.

Because of the following reasons, I choosed the idea 3) in this commit.

The reason why I didn't adopt the idea 1):

The simplest solution is to specify the maximum kernel symbol size for wchan.
But they are 60 characters or so.
The output of ps command become too long and it will be unreadable.

Moreover, some of them are different for the trailing characters
insensibly as follows. So if I consider the optimal size of
wchan, it seems to become long.

   __pci_fixup_resume_earlyasus_hides_smbus_lpc_ich6_resume_early1582
   __pci_fixup_resume_earlyasus_hides_smbus_lpc_ich6_resume_early1583
   __pci_fixup_resume_earlyasus_hides_smbus_lpc_ich6_resume_early1584

The reason why I didn't adopt the idea 2):

Use nwchan instead of wchan to output the address of the kernel
function where the process is sleeping.  sosreport collects
/proc/kallsyms so I can compare the symbol and the address.  But
nwchan seems to be broken because it only output the column size
for 5 or 6 characters. It needs 16 characters on 64bit os.

Signed-off-by: Tomofumi Yoshida <yoshida.tomo@fujitsu.com>